Extract hook usage to dedicated custom hooks
